\name{abmgravity-package}
\alias{abmgravity-package}
\alias{abmgravity}
\docType{package}
\title{
  Functions related to the gravity-like model for ABM
}
\description{
  In ABM, the synthetic populations are created to mimic the real world. Often time, 
  the links among agents are created using a gravity-like model, i.e., the probability
  of a link is proportionate to the "weights" and inversely proportionate to the distance.
  This package has a few functions that do that.
}
\details{
  TBC
}
\author{
  Maintainer: Ben Toh <kokbent@ufl.edu>
}
\references{
  TBC
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
