% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_by_gravity}
\alias{assign_by_gravity}
\title{Assigning ONE location to each point}
\usage{
assign_by_gravity(
  pts,
  locs,
  weights,
  num_loc,
  seed,
  min_x = -87.78555,
  min_y = 24.4699,
  steps = 2L,
  use_capacity = FALSE
)
}
\arguments{
\item{pts}{A matrix of two columns containing xy-coordinates of points (e.g., person or house).}

\item{locs}{A matrix of two columns containing xy-coordinates of locations (e.g., schools or workplaces).}

\item{weights}{A vector of weights for each location (e.g., school or workplace size)}

\item{num_loc}{(Approximate) Number of nearest locations to consider for assignment (to each point).}

\item{seed}{Random number seed.}

\item{min_x}{Smallest y value (helps with grid calculation)}

\item{steps}{Step size when doing neighbourhood searching (affects efficiency)}

\item{use_capacity}{Logical indicating if the weights specified are to be interpreted as capacity; 
If TRUE, each time a location is allocated to a point, the weight/capacity is reduced by one. Default is FALSE.}
}
\value{
A matrix of two columns containing specifying the assignment: row id of point and row id of location.
}
\description{
Assigning ONE location to each point
}
