% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{assign_by_gravity2}
\alias{assign_by_gravity2}
\title{Assigning one or more locations to each point}
\usage{
assign_by_gravity2(
  pts,
  locs,
  weights,
  num_loc_choose,
  num_loc_candidate,
  seed,
  min_x = -87.78555,
  min_y = 24.4699,
  steps = 2L,
  use_capacity = FALSE,
  replace = FALSE
)
}
\arguments{
\item{pts}{A matrix of two columns containing xy-coordinates of points (e.g., person or house).}

\item{locs}{A matrix of two columns containing xy-coordinates of locations (e.g., schools or workplaces).}

\item{weights}{A vector of weights for each location (e.g., school or workplace size)}

\item{num_loc_choose}{Number of locations to be assigned to each point.}

\item{num_loc_candidate}{(Approximate) Number of nearest locations to consider for assignment (to each point).}

\item{seed}{Random number seed.}

\item{min_x}{Smallest y value (helps with grid calculation)}

\item{steps}{Step size when doing neighbourhood searching (affects efficiency)}

\item{use_capacity}{Logical indicating if the weights specified are to be interpreted as capacity; 
If TRUE, each time a location is allocated to a point, the weight/capacity is reduced by one. Default is FALSE.}

\item{replace}{Logical indicating if a location can be assigned to a point multiple times.}
}
\value{
A matrix of two columns containing specifying the assignment: row id of point and row id of location.
}
\description{
Assigning one or more locations to each point
}
\examples{
pts <- matrix(c(-87.45, -87.0, 24.85, 25.05), 2, 2)
locs <- matrix(c(-87.4, -87.5, -87.1, -87.1,
                  24.8,  24.9,  25.1,  25.0), 4, 2),
assign_by_gravity2(pts, locs, c(1, 1, 1, 1), 2, 3, 4328)
}
