% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_network}
\alias{build_network}
\title{Build network among a set of locations}
\usage{
build_network(
  locs,
  weights,
  num_loc_candidate,
  seed,
  min_x = -87.78555,
  min_y = 24.4699,
  steps = 2L
)
}
\arguments{
\item{locs}{A matrix of two columns containing xy-coordinates of locations.}

\item{weights}{A vector of expected number of edges for each location.}

\item{num_loc_candidate}{(Approximate) Number of nearest locations to establish edges with.}

\item{seed}{Random number seed.}

\item{min_x}{Smallest y value (helps with grid calculation)}

\item{steps}{Step size when doing neighbourhood searching}
}
\value{
A matrix of two columns containing the edges: row id of location 1 and location 2.
}
\description{
Build network among a set of locations
}
\examples{
locs <- matrix(c(-87.45, -87.0, -87.4, -87.5, -87.1, -87.1, 
                  24.8,   24.9,  25.1,  25.0,  24.85, 25.05), 6, 2)
build_network(locs, rep(2, 6), 3, 4326)
}
